requirements.txt
chromadb==0.5.0
Flask==3.0.2
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-experimental==0.0.59
langchain-google-genai==1.0.4
langchain-openai==0.1.7
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.26
openpyxl==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
streamlit==1.34.0
tabulate==0.9.0
pandasai==2.1.1